moved attempt_imports to dependencies.py
