+++ tests/neuralnet/test_keras.py
@@ -3,7 +3,7 @@ import pyomo.environ as pyo
 import pytest
 
 from pyomo.common.dependencies import DeferredImportError
-from omlt.io import keras, keras_available
+from omlt.dependencies import keras, keras_available
 if keras_available:
     from omlt.io.keras import load_keras_sequential
 
@@ -13,6 +13,11 @@ from omlt.neuralnet.activations import ComplementarityReLUActivation
 from omlt.scaling import OffsetScaling
 
 from conftest import get_neural_network_data
+
+@pytest.mark.skipif(keras_available, reason="Test only valid when keras not available")
+def test_keras_not_available_exception(datadir):
+    with pytest.raises(DeferredImportError):
+        NN = keras.models.load_model(datadir.file("keras_linear_131_relu"))
 
 @pytest.mark.skipif(not keras_available, reason="Need keras for this test")
 def _test_keras_linear_131(keras_fname, reduced_space=False):
@@ -148,10 +153,6 @@ def test_keras_linear_big(datadir):
 def test_keras_linear_big_reduced_space(datadir):
     _test_keras_linear_big("./models/big", reduced_space=True)
 
-
-def test_keras_not_available_exception(datadir):
-    with pytest.raises(DeferredImportError):
-        NN = keras.models.load_model(datadir.file("keras_linear_131_relu"))
     
 @pytest.mark.skipif(not keras_available, reason="Need keras for this test")
 def test_scaling_NN_block(datadir):
