+++ tests/neuralnet/test_relu.py
@@ -3,7 +3,7 @@ import pyomo.environ as pyo
 import pytest
 
 from omlt.block import OmltBlock
-from omlt.io import onnx_available
+from omlt.dependencies import onnx_available
 from omlt.neuralnet import (
     FullSpaceNNFormulation,
     ReluBigMFormulation,
